# forest-back-demo/docs/integrations.openapi.yaml
openapi: 3.1.0
info:
  title: 산림재난 통합 연계 API
  version: 1.1.0
  description: |
    투비유니콘 통합 플랫폼과 장비·통신망·AI 제공업체 사이의 공통 계약입니다.
    외부 업체는 결과 수신 API를 호출하며, 투비유니콘은 등록된 endpointEnv를 통해
    외부 기능을 호출합니다. 이 명세는 외부 장비·RF·AI 자체 구현을 의미하지 않습니다.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
...
  /assets/{assetId}/credentials:
    get:
      summary: 장치 인증정보 메타데이터 조회
      description: 인증 해시와 원문 secret은 반환하지 않는다.
      operationId: listDeviceCredentials
      parameters:
        - $ref: "#/components/parameters/AssetId"
      responses:
        "200": {description: 발급·만료·폐기·최근 인증 시각 목록}
    post:
      summary: 장치 API 인증정보 발급
      description: 원문 secret은 이 응답에서 한 번만 제공하고 DB에는 해시만 저장한다.
      operationId: issueDeviceCredential
      parameters:
        - $ref: "#/components/parameters/AssetId"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                credentialType: {type: string, const: API_KEY, default: API_KEY}
                expiresAt: {type: string, format: date-time}
      responses:
        "201": {description: 장치 인증정보 발급}
  /assets/{assetId}/credentials/{credentialId}/revoke:
    post:
      summary: 장치 인증정보 폐기
      operationId: revokeDeviceCredential
      parameters:
        - $ref: "#/components/parameters/AssetId"
        - name: credentialId
          in: path
          required: true
          schema: {type: string, format: uuid}
      responses:
        "200": {description: 폐기 완료}
        "404": {description: 인증정보 없음}
  /events/{eventId}/personnel-device-assignments:
    get:
      summary: 사건별 대원-단말 배정 이력 조회
      operationId: listPersonnelDeviceAssignments
      parameters:
        - $ref: "#/components/parameters/EventId"
      responses:
        "200": {description: 배정 이력}
    post:
      summary: 대원에게 사전등록 단말 배정
      operationId: assignDeviceToPersonnel
      parameters:
        - $ref: "#/components/parameters/EventId"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/PersonnelDeviceAssignment"}
      responses:
        "201": {description: 배정 완료}
        "409": {description: 대원 또는 단말이 이미 활성 배정됨}
  /events/{eventId}/personnel-device-assignments/{assignmentId}/release:
    post:
      summary: 대원-단말 배정 해제
      operationId: releasePersonnelDeviceAssignment
      parameters:
        - $ref: "#/components/parameters/EventId"
        - name: assignmentId
          in: path
          required: true
          schema: {type: string, format: uuid}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [releasedAt]
              properties:
                releasedAt: {type: string, format: date-time}
      responses:
        "200": {description: 배정 해제 완료}
  /devices/activate:
    post:
      summary: 현장 단말 인증 및 활성 배정 확인
      security: []
      operationId: activateDevice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [assetId, credential, eventId, personExternalId]
              properties:
                assetId: {type: string, format: uuid}
                credential: {type: string, writeOnly: true}
                eventId: {type: string, format: uuid}
                personExternalId: {type: string}
      responses:
        "200":
          description: 인증 및 배정 확인 완료, 15분 유효 장치 Bearer 토큰 발급
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      activated: {type: boolean, const: true}
                      assetId: {type: string, format: uuid}
                      assignmentId: {type: string, format: uuid}
                      activatedAt: {type: string, format: date-time}
                      accessToken: {type: string, writeOnly: true}
                      tokenType: {type: string, const: Bearer}
                      expiresIn: {type: integer, const: 900}
        "401": {description: 장치 인증 실패}
        "403": {description: 해당 사건·대원에게 배정되지 않은 장치}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    EventId:
      name: eventId
      in: path
      required: true
      schema: {type: string, format: uuid}
...
    AssetId:
      name: assetId
      in: path
      required: true
      schema: {type: string, format: uuid}
  schemas:
...
    PersonnelDeviceAssignment:
      type: object
      required: [personExternalId, assetId, assignedAt]
      properties:
        personExternalId: {type: string, maxLength: 200}
        assetId: {type: string, format: uuid}
        assignedAt: {type: string, format: date-time}
        assignedBy: {type: string, maxLength: 200}
...